requirements.txt
Appium-Python-Client==5.0.0
PySocks==1.7.1
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
future==1.0.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pip==25.0.1
pluggy==1.5.0
pyOpenSSL==25.0.0
pycparser==2.22
pydantic==2.3.0
pydantic-settings==2.0.3
pytest==8.3.5
python-dotenv==1.0.0
selene==2.0.0rc9
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
urllib3-secure-extra==0.1.0
websocket-client==1.8.0
wsproto==1.2.0